+++ packages/edit/client/views/items.js
@@ -1,15 +1,11 @@
 
-
-
 Template.item_place_edit.onRendered(function() {
 	
-	var self = this,
-		tmpl = Template.instance();
+	var self = this;
 
 	self.$('.btn-placedel').btsConfirmButton(function(e) {
-		
 		Meteor.call('removePlace', self.data.id, function(err,res) {
-			K.Map.removeItem(self.data);		
+			K.Map.removeItem(K.placeById(self.data.id));
 			//Router.go('root');
 		});			
 	});