+++ packages/lib/client/L.Cursor.js
@@ -43,7 +43,7 @@ L.Cursor = L.Layer.extend({
 		this._map.removeLayer(this.marker);
 	},
 	setLoc: function(latlng) {
-		this.marker.setLatLng(latlng).addTo(this._map);
+		this.marker.setLatLng(latlng).addTo(this._map).openPopup();
 	}
 });