+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
-...
+- fix all removeItem hide marker
+- disable gps when gone offline
 
+## v1.5.8
 - movable places
 - plugin photos support creation new place by exif photo
 - plugin photos show news photos in sidebar